Week_8_Fine-Tuning/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2
llama-cpp-python==0.2.72
requests==2.31.0
pydantic==2.6.4